[Feature kbss-cvut/fta-fmea-ui#25] Add configuration to disable registration of new user
