+++ src/main/java/cz/cvut/kbss/analysis/controller/AuthController.java
@@ -6,20 +6,15 @@ import cz.cvut.kbss.analysis.dto.authentication.AuthenticationResponse;
 import cz.cvut.kbss.analysis.model.User;
 import cz.cvut.kbss.analysis.service.JwtTokenProvider;
 import cz.cvut.kbss.analysis.service.UserRepositoryService;
-import cz.cvut.kbss.jsonld.JsonLd;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
 
 @RestController
 @RequestMapping("/auth")
@@ -31,15 +26,6 @@ public class AuthController {
     private final JwtTokenProvider jwtTokenProvider;
     private final UserRepositoryService userRepositoryService;
 
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping(value = "/register", consumes = {JsonLd.MEDIA_TYPE}, produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<Void> register(@RequestBody User user) {
-        log.info("> register - {}", user.getUsername());
-        URI uri = userRepositoryService.register(user);
-
-        log.info("< register - {}", uri);
-        return ResponseEntity.created(uri).build();
-    }
 
     @PostMapping("/signin")
     public AuthenticationResponse signIn(@RequestBody AuthenticationRequest data) {
@@ -64,5 +50,4 @@ public class AuthController {
         userRepositoryService.updateCurrent(userUpdate);
         log.info("< updateCurrent - user {} updated", userUpdate.getUri());
     }
-
 }
